SharedVariable in inspector, overwrite in editor
Allows visibility and editing of the SharedVariables' Value field in the
editor.

The OnBeforeSerialize callback has been changed to only overwrite the
runtimeValue when not playing. This was changed because anywhere you
view the shared variable will trigger serialization, which would
overwrite the runtime value. This change keeps the original desired
behavior of the SharedVariable, but stops it from overwriting the value
at runtime. This approach MAY case some issues if SharedVariables are
created at runtime!